Update Makefile and Docker Compose files, modify table columns, and refactor profile handling

+++ postgres/create_tables.sql
@@ -1,5 +1,5 @@
 create table if not exists telegram_users(
-    id int primary key,
+    id BIGINT primary key,
     first_name text not null,
     username text,
     last_name text,
@@ -10,12 +10,12 @@ create table if not exists telegram_users(
 create table if not exists coworking(
     id serial primary key,
     status text not null,
-    responsible_id integer not null references telegram_users(id),
+    responsible_id BIGINT not null references telegram_users(id),
     duration integer default null,
     created_at timestamp not null default NOW()
 );
 create table if not exists profiles(
-    user_id int primary key references telegram_users(id) on delete cascade,
+    user_id BIGINT primary key references telegram_users(id) on delete cascade,
     fio text,
     email text,
     educational_group text,
@@ -27,7 +27,7 @@ create table if not exists profiles(
     company text
 );
 create table if not exists subscriptions(
-    id int primary key references telegram_users(id) on delete cascade,
+    id BIGINT primary key references telegram_users(id) on delete cascade,
     coworking boolean default false,
     hack_club boolean default false,
     design_club boolean default false,
@@ -45,8 +45,8 @@ create table if not exists clubs(
 create table if not exists admin_invite_codes(
     id serial primary key,
     code text,
-    admin_id integer not null REFERENCES telegram_users(id) on DELETE CASCADE,
-    user_id integer REFERENCES telegram_users(id) on DELETE CASCADE,
+    admin_id BIGINT not null REFERENCES telegram_users(id) on DELETE CASCADE,
+    user_id BIGINT REFERENCES telegram_users(id) on DELETE CASCADE,
     created_at TIMESTAMP WITH TIME zone not null default CURRENT_TIMESTAMP,
     activated_at TIMESTAMP WITH TIME zone
 );
